Move action of loading more emails to context
